Move ObjectLabel from Domain to DomainStructure
