+++ src/ControlSystem/ControlErrors/Expansion.hpp
@@ -11,7 +11,7 @@
 #include "ControlSystem/Tags.hpp"
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataVector.hpp"
-#include "Domain/ObjectLabel.hpp"
+#include "Domain/Structure/ObjectLabel.hpp"
 #include "Domain/Tags/ObjectCenter.hpp"
 #include "Options/Options.hpp"
 #include "Parallel/GlobalCache.hpp"
