+++ src/ControlSystem/ControlErrors/Shape.hpp
@@ -11,7 +11,7 @@
 #include "ControlSystem/Protocols/ControlError.hpp"
 #include "ControlSystem/Tags.hpp"
 #include "DataStructures/DataVector.hpp"
-#include "Domain/ObjectLabel.hpp"
+#include "Domain/Structure/ObjectLabel.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/SpherepackIterator.hpp"
 #include "Options/Options.hpp"
 #include "Parallel/GlobalCache.hpp"
