+++ src/Domain/CMakeLists.txt
@@ -21,7 +21,6 @@ spectre_target_sources(
   InterfaceLogicalCoordinates.cpp
   JacobianDiagnostic.cpp
   MinimumGridSpacing.cpp
-  ObjectLabel.cpp
   SizeOfElement.cpp
   SurfaceJacobian.cpp
   Tags.cpp
@@ -48,7 +47,6 @@ spectre_target_headers(
   InterfaceLogicalCoordinates.hpp
   JacobianDiagnostic.hpp
   MinimumGridSpacing.hpp
-  ObjectLabel.hpp
   OptionTags.hpp
   SizeOfElement.hpp
   SurfaceJacobian.hpp
