+++ src/Domain/Structure/ObjectLabel.hpp
