+++ tests/Unit/Domain/CMakeLists.txt
@@ -25,7 +25,6 @@ set(LIBRARY_SOURCES
   Test_InterfaceLogicalCoordinates.cpp
   Test_JacobianDiagnostic.cpp
   Test_MinimumGridSpacing.cpp
-  Test_ObjectLabel.cpp
   Test_Protocols.cpp
   Test_SizeOfElement.cpp
   Test_SurfaceJacobian.cpp
