+++ tests/Unit/Domain/Test_Tags.cpp
@@ -18,10 +18,10 @@
 #include "Domain/Creators/Brick.hpp"
 #include "Domain/Creators/DomainCreator.hpp"
 #include "Domain/ElementMap.hpp"
-#include "Domain/ObjectLabel.hpp"
 #include "Domain/OptionTags.hpp"
 #include "Domain/Protocols/Metavariables.hpp"
 #include "Domain/Structure/ElementId.hpp"
+#include "Domain/Structure/ObjectLabel.hpp"
 #include "Domain/Structure/SegmentId.hpp"
 #include "Domain/Tags.hpp"
 #include "Domain/Tags/ObjectCenter.hpp"
